+++ db/hooks.php
@@ -29,4 +29,8 @@ $callbacks = [
         'hook' => \core\hook\output\before_standard_top_of_body_html_generation::class,
         'callback' => '\local_credentiumclaim\local\hook_callbacks::before_standard_top_of_body_html',
     ],
+    [
+        'hook' => \core_user\hook\extend_user_menu::class,
+        'callback' => '\local_credentiumclaim\local\hook_callbacks::extend_user_menu',
+    ],
 ];